+++ cli/src/core/package_json/package_json_constants.rs
@@ -225,29 +225,29 @@ pub(crate) const EXPRESS_VERSION: &str = "~0.7.8";
 // @forklaunch/hyper-express
 pub(crate) const HYPER_EXPRESS_VERSION: &str = "~0.7.8";
 // @forklaunch/implementation-billing-base
-pub(crate) const BILLING_BASE_VERSION: &str = "~0.4.0";
+pub(crate) const BILLING_BASE_VERSION: &str = "~0.5.0";
 // @forklaunch/implementation-billing-stripe
-pub(crate) const BILLING_STRIPE_VERSION: &str = "~0.1.0";
+pub(crate) const BILLING_STRIPE_VERSION: &str = "~0.2.0";
 // @forklaunch/implementation-iam-base
-pub(crate) const IAM_BASE_VERSION: &str = "~0.4.0";
+pub(crate) const IAM_BASE_VERSION: &str = "~0.5.0";
 // @forklaunch/implementation-worker-bullmq
-pub(crate) const WORKER_BULLMQ_VERSION: &str = "~0.4.0";
+pub(crate) const WORKER_BULLMQ_VERSION: &str = "~0.5.0";
 // @forklaunch/implementation-worker-redis
-pub(crate) const WORKER_REDIS_VERSION: &str = "~0.4.0";
+pub(crate) const WORKER_REDIS_VERSION: &str = "~0.5.0";
 // @forklaunch/implementation-worker-database
-pub(crate) const WORKER_DATABASE_VERSION: &str = "~0.4.1";
+pub(crate) const WORKER_DATABASE_VERSION: &str = "~0.5.0";
 // @forklaunch/implementation-worker-kafka
-pub(crate) const WORKER_KAFKA_VERSION: &str = "~0.4.0";
+pub(crate) const WORKER_KAFKA_VERSION: &str = "~0.5.0";
 // @forklaunch/infrastructure-redis
 pub(crate) const INFRASTRUCTURE_REDIS_VERSION: &str = "~0.2.8";
 // @forklaunch/infrastructure-s3
 pub(crate) const INFRASTRUCTURE_S3_VERSION: &str = "~0.2.9";
 // @forklaunch/interfaces-billing
-pub(crate) const BILLING_INTERFACES_VERSION: &str = "~0.4.0";
+pub(crate) const BILLING_INTERFACES_VERSION: &str = "~0.5.0";
 // @forklaunch/interfaces-iam
-pub(crate) const IAM_INTERFACES_VERSION: &str = "~0.4.0";
+pub(crate) const IAM_INTERFACES_VERSION: &str = "~0.5.0";
 // @forklaunch/interfaces-worker
-pub(crate) const WORKER_INTERFACES_VERSION: &str = "~0.3.0";
+pub(crate) const WORKER_INTERFACES_VERSION: &str = "~0.4.0";
 // @forklaunch/internal
 pub(crate) const INTERNAL_VERSION: &str = "~0.1.9";
 // @forklaunch/universal-sdk