+++ blueprint/implementations/worker/redis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/implementation-billing-base
 
+## 0.5.0
+
+### Minor Changes
+
+- pin forklaunch package to patch versions
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/interfaces-worker@0.4.0
+
 ## 0.4.0
 
 ### Minor Changes